1st upload
DB.sql is the sql structure of this project
Sending format.txt is a self note. telling me how the data to and from
the main server are written and the meaning.

CarPark_MainServer is a socket server that listens to a port,no GUI.

CarPark_Simulator is a simulator that simulates an actual car park.Block
A to H are the blocks name like we see in actual car parks.Each block
have 8 spaces.
Green means the space is empty, red means theres a car parked
inside.When clicked they send data to CarPark_MainServer telling that
the state of the space has change.And the mainserver invokes the
CarPark_Live class
The entrance and exit button are the entrance and exit to the parking
lot. when clicked they are processed by the CarPark_Sql class.

CarPark_AdminViewer is a GUI live view on the car park.
currently i'm just viewing my own simulator.
by default, every 10sec we send a request to the server for an update.
the update time rate is changeable.